backend/requirements.txt
fastapi
uvicorn
pydantic
pinecone
SQLAlchemy
python-dotenv
psycopg2-binary
voyageai